# packages/contracts/openapi/kiwi-agent-tools-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Kiwi Agent Tools API
  version: 1.0.0-rc.1
  description: >-
    Contrato versionado para agentes externos. Las escrituras crean propuestas o borradores;
    nunca confirman compras ni aplican movimientos de inventario.
servers:
  - url: https://{kiwiHost}
    variables:
      kiwiHost:
        default: kiwi.example.invalid
security:
  - serviceBearer: []
...
paths:
...
  /api/v1/agent-tools/operations/{operation_id}:
    get:
      tags: [Agent operations]
      operationId: getAgentOperation
      summary: Recuperar estado autoritativo de una operación propia
      parameters:
        - name: operation_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Estado autoritativo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentOperation"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
components:
  securitySchemes:
    clientCredentials:
      type: http
      scheme: basic
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
...
  responses:
...
    InvalidRequest:
      description: Esquema o referencia inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BusinessError"
...
    Unauthorized:
      description: Credencial ausente, inválida o vencida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BusinessError"
    Forbidden:
      description: Integración, capacidad o sucursal no autorizada
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BusinessError"
    NotFound:
      description: Operación no encontrada dentro de la identidad autorizada
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BusinessError"
...
    RateLimited:
      description: Límite acotado excedido
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BusinessError"
    DependencyUnavailable:
      description: Dependencia temporalmente no disponible; no implica éxito
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BusinessError"
  schemas:
    Uuid:
      type: string
      format: uuid
...
    OperationStatus:
      type: string
      enum:
        - RECEIVED
        - VALIDATED
        - READY_FOR_REVIEW
        - DRAFT_CREATED
        - REJECTED
        - APPLIED
        - EXPIRED
        - CONFIRMED_BY_HUMAN
        - CANCELLED_BY_HUMAN
    AgentReasonCode:
      type: string
      enum:
        - validation_rejected
        - stale_reference
        - human_rejected
        - expired
        - canonical_apply_failed
        - canonical_confirmed
        - canonical_cancelled
        - callback_destination_denied
        - callback_delivery_exhausted
    AgentOperation:
      type: object
      additionalProperties: false
      required: [operation_id, operation_type, status, created_at, updated_at]
      properties:
        operation_id:
          $ref: "#/components/schemas/Uuid"
        operation_type:
          type: string
          enum: [catalog_proposal, inventory_item_proposal, recipe_proposal, purchase_draft]
        status:
          $ref: "#/components/schemas/OperationStatus"
        resource_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        reason_code:
          oneOf:
            - $ref: "#/components/schemas/AgentReasonCode"
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BusinessError:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          $ref: "#/components/schemas/BusinessErrorDetail"
    BusinessErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id]
      properties:
        code:
          type: string
          enum:
            - permission_denied
            - agent_unauthorized
            - agent_disabled
            - agent_capability_denied
            - agent_branch_denied
            - agent_identity_not_found
            - agent_integration_not_configured
            - agent_schema_invalid
            - callback_destination_denied
            - idempotency_conflict
            - stale_reference
            - rate_limited
            - dependency_unavailable
            - operation_not_found
        message:
          type: string
          maxLength: 300
        correlation_id:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 100
            - type: "null"